Rev 0.4
Version ready for detailed testing. This version calibrates the background, processes target status to filter out bad data, puts negative codes in the adjusted data array for bad data, out of range data, and calibration data.  The processed range values are printed out and the smallest value and its coordinates are printed with each measurement pass.  Measurements are slowed down for close observation of the data,